requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cycler==0.10.0
distlib==0.3.0
Django==3.0.7
django-channels==0.7.0
filelock==3.0.12
Flask==1.1.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
msgpack==0.6.2
numpy==1.18.5
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
txaio==20.4.1
urllib3==1.25.9
virtualenv==20.0.20
Werkzeug==1.0.1
zope.interface==5.1.0
